Refactor: relocate verifier classes
