+++ src/main/java/edu/tamu/di/SAFCreator/model/Bitstream.java
@@ -36,7 +36,7 @@ import org.apache.tika.mime.MimeTypeException;
 import org.apache.tika.mime.MimeTypes;
 
 import edu.tamu.di.SAFCreator.enums.FlagColumns;
-import edu.tamu.di.SAFCreator.model.Verifier.Problem;
+import edu.tamu.di.SAFCreator.model.verify.Verifier.Problem;
 
 public class Bitstream extends CellDatumImpl {
     private static int TimeoutRead = 20000;
