+++ src/main/java/edu/tamu/di/SAFCreator/model/Item.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import edu.tamu.di.SAFCreator.Util;
-import edu.tamu.di.SAFCreator.model.Verifier.Problem;
+import edu.tamu.di.SAFCreator.model.verify.Verifier.Problem;
 
 public class Item {
     private Batch batch;
