+++ src/main/java/edu/tamu/di/SAFCreator/model/verify/VerifierProperty.java
@@ -1,4 +1,4 @@
-package edu.tamu.di.SAFCreator.model;
+package edu.tamu.di.SAFCreator.model.verify;
 
 public interface VerifierProperty {
     public boolean generatesError();
